Add ApplicationTags and HeartbeaterService; Rename everything to Wavefront.SDK.CSharp; Enable WavefrontHistogram for ASP.NET Core SDK
