--- Wavefront.CSharp.SDK/Entities/Tracing/SpanLog.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace Wavefront.CSharp.SDK.Entities.Tracing
-{
-    /// <summary>
-    /// SpanLog defined as per the opentracing.io specification
-    /// </summary>
-    public class SpanLog
-    {
-        private readonly long timestamp;
-        private readonly IDictionary<string, string> fields;
-
-        public SpanLog(long timestamp, IDictionary<string, string> fields)
-        {
-            this.timestamp = timestamp;
-            this.fields = fields;
-        }
-    }
-}
