+++ Wavefront.SDK.CSharp.sln
@@ -1,0 +1,37 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wavefront.SDK.CSharp", "Wavefront.SDK.CSharp\Wavefront.SDK.CSharp.csproj", "{0CC170B1-8B06-403C-817E-CEC653B67DEC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Wavefront.SDK.CSharp.Test", "Wavefront.SDK.CSharp.Test\Wavefront.SDK.CSharp.Test.csproj", "{2843F32B-90B0-44AE-87D9-F1BD3442A7C6}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{456CE097-21CF-45D7-B07D-BFDDF096D19D}"
+	ProjectSection(SolutionItems) = preProject
+		README.md = README.md
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0CC170B1-8B06-403C-817E-CEC653B67DEC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0CC170B1-8B06-403C-817E-CEC653B67DEC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0CC170B1-8B06-403C-817E-CEC653B67DEC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0CC170B1-8B06-403C-817E-CEC653B67DEC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2843F32B-90B0-44AE-87D9-F1BD3442A7C6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2843F32B-90B0-44AE-87D9-F1BD3442A7C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2843F32B-90B0-44AE-87D9-F1BD3442A7C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2843F32B-90B0-44AE-87D9-F1BD3442A7C6}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(MonoDevelopProperties) = preSolution
+		Policies = $0
+		$0.TextStylePolicy = $1
+		$1.TabsToSpaces = True
+		$1.scope = text/x-csharp
+		$1.FileWidth = 100
+		$0.CSharpFormattingPolicy = $2
+		$2.scope = text/x-csharp
+	EndGlobalSection
+EndGlobal
