+++ Wavefront.SDK.CSharp/Wavefront.SDK.CSharp.csproj
@@ -1,0 +1,29 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <PackOnBuild>true</PackOnBuild>
+    <PackageId>Wavefront.SDK.CSharp</PackageId>
+    <PackageVersion>0.1.0</PackageVersion>
+    <Authors>Wavefront</Authors>
+    <Description>This package provides support for sending metrics, histograms and opentracing spans to Wavefront via proxy or direct ingestion.</Description>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Folder Include="Proxy\" />
+    <Folder Include="Common\" />
+    <Folder Include="Entities\" />
+    <Folder Include="Entities\Histograms\" />
+    <Folder Include="Entities\Metrics\" />
+    <Folder Include="Entities\Tracing\" />
+    <Folder Include="Entities\Events\" />
+    <Folder Include="DirectIngestion\" />
+    <Folder Include="Common\Application\" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="2.1.1" />
+    <PackageReference Include="NuGet.Build.Packaging" Version="0.2.2" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="2.1.1" />
+    <PackageReference Include="TDigest" Version="1.0.6" />
+  </ItemGroup>
+</Project>
